+++ android/main/MainActivity.java
@@ -1,5 +1,7 @@
 package com.wenathlan.saddle;
 
+package com.wenathlan.saddle;
+
 import com.getcapacitor.BridgeActivity;
 
 public class MainActivity extends BridgeActivity {}